refactor: consolidate time-aware projection into deep module scenarioProjection
Why:
Unit state projection, hierarchy projection, and feature-layer state
projection were split across time.ts and hierarchy.ts with many leaky
internals exported (updateCurrentUnitState, syncTimedHierarchyProjection,
refreshHierarchyTimelineMetadata, etc.). Callers in 6+ files had to
manually coordinate reactive counters (unitStateCounter, isMapStylesDirty)
and style cache invalidation, which made the surface area wide and the
system hard to reason about.

What:
- Create scenarioProjection.ts — a deep module that hides all projection
  complexity behind 5 public functions:
  - projectScenarioToTime(state, timestamp)
  - reprojectUnit(state, unitId)
  - reprojectHierarchy(state, options?)
  - refreshHierarchyTimeline(state)
  - createInitialState(unit)
- Strip projection logic from time.ts (432 -> 217 lines); it now only
  handles time arithmetic and event management.
- Delete hierarchy.ts and its tests after moving logic into the new module.
- Update all callers (unitStateManipulations, toeManipulations,
  supplyManipulations, unitManipulations, settingsManipulations, io.test)
  to use the narrow public interface instead of raw primitives.
- Merge time.test.ts and hierarchy.test.ts into
  scenarioProjection.test.ts, rewritten against the public API.

Result: fewer exports, hidden side-effect coordination, simpler mental
model for consumers.
