+++ src/scenariostore/unitManipulations.ts
@@ -17,7 +17,6 @@ import { CUSTOM_SYMBOL_SID_INDEX, SID_INDEX, Sidc } from "@/symbology/sidc";
 import { setCharAt } from "@/components/helpers";
 import { SID } from "@/symbology/values";
 import { klona } from "klona";
-import { createInitialState } from "@/scenariostore/time";
 import { computed } from "vue";
 import type {
   Unit,
@@ -37,9 +36,10 @@ import {
 } from "@/scenariostore/unitStateManipulations";
 import { CUSTOM_SYMBOL_PREFIX } from "@/config/constants.ts";
 import {
-  refreshHierarchyTimelineMetadata,
-  syncTimedHierarchyProjection,
-} from "@/scenariostore/hierarchy";
+  createInitialState,
+  reprojectHierarchy,
+  refreshHierarchyTimeline,
+} from "@/scenariostore/scenarioProjection";
 
 export type NWalkSubUnitCallback = (unit: NUnit) => void;
 
@@ -144,7 +144,7 @@ export function useUnitManipulations(store: NewScenarioStore) {
   } = useUnitStateManipulations(store);
 
   function refreshProjectedHierarchy(s = state) {
-    syncTimedHierarchyProjection(s, s.currentTime, { force: true });
+    reprojectHierarchy(s, { force: true });
   }
 
   function removeTimedHierarchyReferences(targetIds: Set<EntityId>, s = state) {
@@ -164,7 +164,7 @@ export function useUnitManipulations(store: NewScenarioStore) {
       }
     });
     if (anyChanged) {
-      refreshHierarchyTimelineMetadata(s);
+      refreshHierarchyTimeline(s);
     }
   }
